requirements.txt
fastapi
uvicorn
langchain-community
sentence-transformers
faiss-cpu
langchain-google-genai
python-dotenv

